変数管理を分割
